Redoes how turfs share with each other
Ok so turfs used to pass in "only" their atmos adjacent turfs into share()
This causes problems, but to understand that you first need to understand how turfs share with each other

Each pair of turfs will only ever call share() on each other once. They use an archived cycle to keep track of
this ordering

That means turf A calling share on turf B should work the same as turf B calling share on turf A
But it doesn't.

Why also requires some more context.

The key logic of FEA, the core sharing system we use is that neighboring cells should equalize as best they can.
So taken on a 2d plane, you'd have two sharing partners, the cells to your left and right

We want to equalize with them as best we can, which means we effectively want to average out our mix with
theirs. There's an equation for this which we do not use because it means each pair of turfs needs to talk to
each other twice, which is pain expensive. That and I'm pretty sure it would prevent us from yielding

So instead of a complex form of averaging, we portion up tiles. So if you have two neighbors and you have
something they don't, you can give them each a third. Have to keep one for ourselves mind, because otherwise
we'll run out of gas.

The math for this looks like (totaldeltagas)/(neighborcount + 1)

You may notice this in the current share(). it has a flaw though. A new flaw.

It used to be that this number was hardcoded to 5. This means that turf A -> turf B is the same as turf B ->
turf A, because they're each portioning up the gas in the same way.

Since surely none would be fucking insane enough to add new levels of space you can share wi- oh god multiz.

So when multiz happen we went from a hardcoded neighbor count to a variable one, depending on the amount of
adjacent turfs. I hate kevinz.

Because of this, turf A -> turf B was no longer the same as turf B -> turf A if one of those turfs had
a different neighbor count, from I DON'T KNOW WALLS?

The fix for this is to use our neighbor count when moving gas from our tile to someone elses, and use the
sharer's neighbor count when taking from it.

This makes sense intuitively if you think of it like portioning up a tile, but I've included a rundown to make
it a bit easier to prove to yourself.

START

I have 10
You have 20
let's share
I've got 2 partners
you've got 3 partners
so you want to give me 1/4th of your gas
I want to give you 1/3rd of my gas

the total gas diff between me and you is -10
since it's negative you get to decide how to portion it
so the total amount to share is -2.5
I end up with 12.5
you end up with 17.5

again

total diff is -5
to share is 1.25
I end up with 13.75
you end up with 16.25

again

total diff is -2.5
to share is 0.3125
I end up with 14.0625
you end up with 15.9375

END

anywayyyyyy that's what I'm doing here. I didn't really get into the why, but basically right now we need to
check for negative moles each share() in both mixes, which is expensive and a pain. I'd like to just make
the system logically complete so we don't need to do that unless we're actually creating a new gas.

I've also done some of what I think are micro ops to how we assert/add gas, since we can use these lists for
checking if we should garbage collect anyway. I think this is cheap, but honestly I'm not totally sure. just
kinda going off gut ya know?