Adds a proper implementation of "Unfloatify"
Here we go a fucking gain.

Basically, rather then asigning lower z levels their own planes manually
We're going to use float planes to do it automatically.

The idea of float planes is to allow obj a with FLOAT + say, lighting
nested inside object b to inherit the plane of b.

So obj a's plane looks like lighting + b plane

If float planes couldn't overlap this would be easy, but they
unfortunately can.

So if we want turf lighting planes to be properly shifted "down" by an
openspace plane, we're going to need to make sure they take into account
the plane of the thing they're "in"

This also means we can't reuse overlays quite as readily.

Life is pain.