Fixes slow jps resolutions
The pathfinding subsystem runs on default wait, IE 20ds. This is stupid,
and adds random delay to path resolution.
I could make it ticker but that seems wrong. I just moved it to 0.5 so
it matches tick lag

This brought resolution times from 10ds to 1, but it can still go lower.
As things currently work, jps move loops sleep to await updates from the
pathfinder subsystem
This is fine for most uses, but it's not needed here. Can be replaced
with a callback, and it'll be a tick faster since we don't need to wait
until the next tick for the sleep to wear off.
It's faster too, since sleeps have overhead and can clog the MC