Attempts to implement base lighting, light height setters, and delayed processing
Idea of base lighting is it provides a way for multiple sources to fake
being one, really potent source. Useful for starlight or faking area
lighting in a way that doesn't flush everything else out

It needs work done to make it fast enough to be usable. I have a
theoretical implementation written out that would make it index based
and do away with the in-hotcode sorting, but that will have to wait for
a little bit

Light height setting is simple. Light height describes where we are
physically. Most of the time you want that to be at "wall hang" level,
since it makes sense. But sometimes you are the floor, or worse, you
want to display lights at full vivid levels to your 4 corners (IE: you
are making direct contact with your corners). So we should support
sources having different height preferences.

This fixes starlight looking permenantly washed out

Delayed processing of light corners is done to avoid the stupid amount
of corner updates that modifying starlight can cause, much of which
happen to space, which literally does not display light object
modifications.

Because of this, it is better to just not, until the turf becomes not
space. We do this by having space override our "gen light corners" proc
and return the bits of its content that aren't doing anything
interesting as inactive, then we update them later on request.

This is a fineish idea (though it seems like corner and object work
takes WAY less time then just working with sources, so perhaps not?
unclear.

It's just it currently causes hard deletes, and I need to track that
down before I do anything with it. So it stays here for now