Adds lighting height control (space color FIXED)
Adds support for modiying a light's "height"
You can think of this as the distance it is from the ground below it
(Really it's the distance to the corners around it + 0.5 but yaknow)
We use it to keep wall lights from looking weird, but well, not
everything is a wall light

Floors tend to not be, and space in particular does not want to be
treated as such.
In fact, it wants a NEGATIVE height, so it acts as if it in on top of
all of its corners. This allows us to ensure that the starlight from
space and the starlight from starlight overlays have the same intensity
and color, preventing weird lines from where the two intersect