Implements storage pip display in tooltips
This required some fenangling, basically tooltips need to use legacy
assets. We pregenerate a spritesheet with all possible sizes of pips,
and then just use the classname in a span in the ui. Ignoring the pain
of icon work and getting the assets SENT, it was pretty trivial.

Also sends the asset to tgui chat, for later use