Refactors status effects to not CHANGE WHAT VARS MEAN
I need access to/the ability to modify the durations of status effects,
but they change what the vars MEAN after init, which is STUPID and I
hate it.

I've added a few helpers that exist to aid in working with the defines
and such, I haven't implemented them globally, can if requested.

This incidentally fixes a few bugs that assumed duration/tick_interval
meant what their names implied (multiplying duration == more effect
later in the round, that sort)